// frontend/src/app/contact/interfaces/contact.d.ts
export interface ContactFormProps {
  email: string;
  message: string;
}
export interface FormDataProps {
  htmlTag: string;
  labelAriaLabel: string;
  labelText: string;
  type?: string;
  id: string;
  name: string;
  placeholder: string;
  inputAriaLabel: string;
  rows?: number;
  maxLength?: number;
}